Ensure ratios are integers
per_page parameter was not working as parameters would get passed to
Ratios as strings, which means we'd always fail to calculte the offset.
Tests were passing because all of them used the first page, which uses
offset 0 independently of the configured ratios. With strings, and any
other page > 1, a `TypeError: String can't be coerced into Integer`
would be raised as we'd try to perform arithmetic operations on strings.